backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pgeocode
firebase-admin
requests
google-generativeai
